requirements.txt
flask
groq
python-telegram-bot==20.7
gunicorn